refactor(gfn): modularize NVST RTSP probe
